記事ページにもoutlet機能追加
